# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DIU QuestionBank API",
    "version": "1.0.0",
    "description": "Backend API for **DIU QuestionBank**, running on Cloudflare Workers.\n\n## Authentication\n\nSign in with a verified `@diu.edu.bd` Google account. Obtain a Google ID token client-side\n(Google Sign-In), then exchange it for a DIU QuestionBank JWT:\n\n```http\nPOST /auth/google\nContent-Type: application/json\n\n{ \"idToken\": \"<google-id-token>\" }\n```\n\nThe response includes a JWT — pass it on every authenticated request:\n\n```\nAuthorization: Bearer <token>\n```\n\nIf the verified DIU email already exists, the existing user is signed in (`200`). If not, a new\nuser is created with an opaque generated username (`201`). Its role is `\"admin\"` when the\nemail matches the configured `ADMIN_EMAIL`; otherwise it is `\"user\"`. Google accounts outside\nthe `diu.edu.bd` domain are rejected (`403`).\n\n## Conventions\n\n- All timestamps are **Unix epoch seconds** (UTC integers), not ISO 8601 strings.\n- Errors return `{ \"error\": \"<message>\" }`. Validation failures additionally include an\n  `issues` array with field-level details."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT obtained from `POST /auth/google`."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "issues": {
            "description": "Field-level details, present on validation failures.",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              },
              "required": [
                "field",
                "message"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "QuestionSubmissions": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991
                },
                "section": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "batch": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "fileSize": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991,
                  "description": "Size of the PDF in bytes."
                },
                "watermarkStatus": {
                  "type": "string",
                  "enum": [
                    "awaiting",
                    "completed",
                    "failed"
                  ]
                },
                "createdAt": {
                  "type": "integer",
                  "minimum": -9007199254740991,
                  "maximum": 9007199254740991,
                  "description": "Unix epoch seconds (UTC)"
                },
                "pdfUrl": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "null"
                    }
                  ],
                  "description": "Absolute URL to the submission PDF, served by `GET /files/:key`."
                },
                "contributor": {
                  "anyOf": [
                    {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "minimum": -9007199254740991,
                          "maximum": 9007199254740991
                        },
                        "name": {
                          "type": "string"
                        },
                        "username": {
                          "type": "string"
                        },
                        "image": {
                          "anyOf": [
                            {
                              "type": "string"
                            },
                            {
                              "type": "null"
                            }
                          ]
                        }
                      },
                      "required": [
                        "id",
                        "name",
                        "username",
                        "image"
                      ],
                      "additionalProperties": false
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "id",
                "section",
                "batch",
                "fileSize",
                "watermarkStatus",
                "createdAt",
                "pdfUrl",
                "contributor"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "data"
        ],
        "additionalProperties": false
      },
...
      "UpdateSubmission": {
        "$schema": "https://json-schema.org/draft/2020-12/schema",
        "type": "object",
        "properties": {
          "questionId": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "userId": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 9007199254740991
          },
          "section": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ]
          },
          "batch": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ]
          },
          "watermarkStatus": {
            "type": "string",
            "enum": [
              "awaiting",
              "completed",
              "failed"
            ]
          }
        },
        "additionalProperties": false
      },
...
      "SubmissionCreateForm": {
        "type": "object",
        "properties": {
          "pdf": {
            "type": "string",
            "format": "binary",
            "description": "PDF file (max 20 MB)."
          },
          "questionId": {
            "type": "integer",
            "description": "Parent question id."
          },
          "userId": {
            "type": "integer",
            "description": "Contributor (user) id."
          },
          "section": {
            "type": "string",
            "description": "Optional section label."
          },
          "batch": {
            "type": "string",
            "description": "Optional batch label."
          }
        },
        "required": [
          "pdf",
          "questionId",
          "userId"
        ]
      },
      "PdfUploadForm": {
        "type": "object",
        "properties": {
          "pdf": {
            "type": "string",
            "format": "binary",
            "description": "Replacement PDF file (max 20 MB)."
          }
        },
        "required": [
          "pdf"
        ]
      },
...
    }
  },
  "paths": {
...
    "/contributors/{username}/submissions": {
      "get": {
        "tags": [
          "contributors"
        ],
        "summary": "List a contributor's submissions",
        "x-badges": [
          {
            "name": "Public",
            "color": "#16a34a",
            "position": "after"
          }
        ],
        "description": "**Access:** `Public` — No authentication required.\n\nA contributor's own submissions, newest first. Paginated. Each row carries its parent question (`id` + `title`) instead of the contributor, and links to its PDF via `pdfUrl` (served by `GET /files/:key`).",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The contributor's username."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "1-based page number."
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Items per page (max 100)."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContributorSubmissionList"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/questions/{id}/submissions": {
      "get": {
        "tags": [
          "questions"
        ],
        "summary": "List a question's submissions",
        "x-badges": [
          {
            "name": "Public",
            "color": "#16a34a",
            "position": "after"
          }
        ],
        "description": "**Access:** `Public` — No authentication required.\n\n**All** submissions for a question (no pagination — a single question has few). Each submission links to its PDF via `pdfUrl` (served by `GET /files/:key`) and includes the contributor, if any.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Question id."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionSubmissions"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/submissions": {
      "get": {
        "tags": [
          "admin-submissions"
        ],
        "summary": "List submissions",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-badges": [
          {
            "name": "Admin",
            "color": "#dc2626",
            "position": "after"
          }
        ],
        "description": "**Access:** `Admin` — Requires a bearer token from an account with `role: \"admin\"`.\n\nPaginated, filterable by `questionId`, `userId`, and `watermarkStatus`. Newest first.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "1-based page number."
          },
          {
            "name": "perPage",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "description": "Items per page (max 100)."
          },
          {
            "name": "questionId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Filter by question id."
          },
          {
            "name": "userId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Filter by contributor (user) id."
          },
          {
            "name": "watermarkStatus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "awaiting",
                "completed",
                "failed"
              ]
            },
            "description": "Filter by watermark status."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSubmissionList"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "admin-submissions"
        ],
        "summary": "Create a submission",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-badges": [
          {
            "name": "Admin",
            "color": "#dc2626",
            "position": "after"
          }
        ],
        "description": "**Access:** `Admin` — Requires a bearer token from an account with `role: \"admin\"`.\n\nUploads a PDF (multipart field `pdf`, max 20 MB) and creates a submission for the given `questionId`, attributed to a `userId` (contributor, required).",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionCreateForm"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSubmission"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "PDF exceeds the 20 MB size limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/submissions/{id}": {
      "get": {
        "tags": [
          "admin-submissions"
        ],
        "summary": "Get a submission by id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-badges": [
          {
            "name": "Admin",
            "color": "#dc2626",
            "position": "after"
          }
        ],
        "description": "**Access:** `Admin` — Requires a bearer token from an account with `role: \"admin\"`.\n\nA single submission with its question, contributor, and file URLs.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Submission id."
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSubmission"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "admin-submissions"
        ],
        "summary": "Update submission metadata",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-badges": [
          {
            "name": "Admin",
            "color": "#dc2626",
            "position": "after"
          }
        ],
        "description": "**Access:** `Admin` — Requires a bearer token from an account with `role: \"admin\"`.\n\nUpdates submission metadata (`questionId`, `userId`, `section`, `batch`, `watermarkStatus`). `section`/`batch` accept `null` to clear them. Use `PUT /admin/submissions/{id}/pdf` to replace the file.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Submission id."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSubmission"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSubmission"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed, or a referenced entity does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "admin-submissions"
        ],
        "summary": "Delete a submission",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-badges": [
          {
            "name": "Admin",
            "color": "#dc2626",
            "position": "after"
          }
        ],
        "description": "**Access:** `Admin` — Requires a bearer token from an account with `role: \"admin\"`.\n\nDeletes a submission and removes its PDF (and watermarked PDF) from storage.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Submission id."
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted — no content"
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "An approved manual submission references this submission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/admin/submissions/{id}/pdf": {
      "put": {
        "tags": [
          "admin-submissions"
        ],
        "summary": "Replace a submission's PDF",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-badges": [
          {
            "name": "Admin",
            "color": "#dc2626",
            "position": "after"
          }
        ],
        "description": "**Access:** `Admin` — Requires a bearer token from an account with `role: \"admin\"`.\n\nReplaces the PDF (multipart field `pdf`, max 20 MB). Resets watermarking to `awaiting` and clears any existing watermarked file.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "Submission id."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/PdfUploadForm"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Replaced",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminSubmission"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed or bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Admin access required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "413": {
            "description": "PDF exceeds the 20 MB size limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}